tasks/6_writing_robust_readable_and_maintainable_code/day_3_documentation_and_security/security/hackme1/app/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-decouple==3.6
